requirements.txt
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0

boto3==1.34.44
botocore==1.34.44
django-storages==1.14.2




Django==4.2.10
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1

django-simple-history==3.3.0
django-import-export==3.2.0


django-cors-headers==4.3.1
django-environ==0.11.2

djoser==2.2.2

pillow==10.2.0
psycopg2==2.9.9
PyJWT==2.8.0
whitenoise==6.6.0

django-countries==7.5.1
twilio==9.0.1